Reimplement unamed-task rule as name[missing]
This reimplementation will allow us to implement closely related
extra checks.

Related: #2171 #2170 #2169 #2036
