+++ tox.ini
@@ -126,7 +126,7 @@ deps =
 setenv =
   PIP_CONSTRAINT = {toxinidir}/requirements.txt
 commands =
-  sh -c "rm -f docs/pkg/*.rst"
+  sh -c "rm -f {toxinidir}/docs/pkg/*.rst"
   {envpython} -m sphinx \
     -j auto \
     -b linkcheck \
